Attempt to fix running test in CI again
