+++ package.json
@@ -26,8 +26,8 @@
     "test:karma:watch": "karma start karma.conf.js --no-single-run",
     "test:karma:prod": "cross-env MINIFY=true NODE_ENV=production karma start karma.conf.js --single-run",
     "test:karma:prod:watch": "cross-env NODE_ENV=production karma start karma.conf.js --no-single-run",
-    "test:mocha": "cross-env COVERAGE=true mocha --require test/node/setup.js --recursive packages/*/test/node/**.test.tsx --recursive packages/react/runtime/test/node/**.test.tsx",
-    "test:mocha:prod": "cross-env COVERAGE=true NODE_ENV=production mocha --require test/node/setup.js --recursive packages/*/test/node/**.test.tsx --recursive packages/react/runtime/test/node/**.test.tsx",
+    "test:mocha": "cross-env COVERAGE=true mocha --require test/node/setup.js --recursive packages/*/test/node/**.test.tsx",
+    "test:mocha:prod": "cross-env COVERAGE=true NODE_ENV=production mocha --require test/node/setup.js --recursive packages/*/test/node/**.test.tsx",
     "docs:start": "cd docs && pnpm start",
     "docs:build": "cd docs && pnpm build",
     "docs:preview": "cd docs && pnpm preview",
