+++ packages/react/runtime/test/node/renderToStaticMarkup.test.tsx
@@ -1,5 +1,4 @@
-import { signal, useSignalEffect } from "../../../";
-import { useSignals } from "./../../../runtime";
+import { signal, useSignalEffect } from "@preact/signals-react";
 import { createElement } from "react";
 import { renderToStaticMarkup } from "react-dom/server";
 import { mountSignalsTests } from "../../../test/shared/mounting";
@@ -21,27 +20,5 @@ describe("@preact/signals-react/runtime", () => {
 			expect(html).to.equal("<p>foo</p>");
 			expect(spy.called).to.be.false;
 		});
-	});
-
-	it("should clean up signal dependencies after executing", async () => {
-		const getTargets = (s: any): any => s.t ?? s._targets;
-
-		const sig = signal(0);
-		function App() {
-			const effectStore = useSignals(/* MANAGED_COMPONENT */ 1);
-			try {
-				return <p>{sig.value}</p>;
-			} finally {
-				effectStore.f();
-			}
-		}
-
-		expect(getTargets(sig)).to.be.undefined;
-
-		const html = renderToStaticMarkup(<App />);
-		expect(html).to.equal("<p>0</p>");
-
-		await Promise.resolve();
-		expect(getTargets(sig)).to.be.undefined;
 	});
 });
