+++ packages/react/test/node/renderToStaticMarkup.test.tsx
@@ -1,4 +1,5 @@
 import { signal, useSignalEffect } from "@preact/signals-react";
+import { useSignals } from "@preact/signals-react/runtime";
 import { createElement } from "react";
 import { renderToStaticMarkup } from "react-dom/server";
 import { mountSignalsTests } from "../shared/mounting";
@@ -18,5 +19,27 @@ describe("renderToStaticMarkup", () => {
 		const html = await renderToStaticMarkup(<App />);
 		expect(html).to.equal("<p>foo</p>");
 		expect(spy.called).to.be.false;
+	});
+
+	it("should clean up signal dependencies after executing", async () => {
+		const getTargets = (s: any): any => s.t ?? s._targets;
+
+		const sig = signal(0);
+		function App() {
+			const effectStore = useSignals(/* MANAGED_COMPONENT */ 1);
+			try {
+				return <p>{sig.value}</p>;
+			} finally {
+				effectStore.f();
+			}
+		}
+
+		expect(getTargets(sig)).to.be.undefined;
+
+		const html = renderToStaticMarkup(<App />);
+		expect(html).to.equal("<p>0</p>");
+
+		await Promise.resolve();
+		expect(getTargets(sig)).to.be.undefined;
 	});
 });
